Rename marks file
