+++ addon/commands/add-mark-command.ts
@@ -1,5 +1,5 @@
 import Command from '@lblod/ember-rdfa-editor/commands/command';
-import {Serializable} from '@lblod/ember-rdfa-editor/model/markSpec';
+import {Serializable} from '@lblod/ember-rdfa-editor/model/mark';
 import ModelRange from '@lblod/ember-rdfa-editor/model/model-range';
 import Model from '@lblod/ember-rdfa-editor/model/model';
 import {ModelError} from '@lblod/ember-rdfa-editor/utils/errors';
