+++ addon/model/model-text.ts
@@ -5,7 +5,7 @@ import ModelNode, {
 import { ModelError } from '@lblod/ember-rdfa-editor/utils/errors';
 import { stringToVisibleText } from '@lblod/ember-rdfa-editor/editor/utils';
 import ModelNodeUtils from '@lblod/ember-rdfa-editor/model/util/model-node-utils';
-import { Mark, MarkSet } from '@lblod/ember-rdfa-editor/model/markSpec';
+import { Mark, MarkSet } from '@lblod/ember-rdfa-editor/model/mark';
 
 const NON_BREAKING_SPACE = '\u00A0';
