+++ addon/model/readers/xml-text-reader.ts
@@ -3,7 +3,7 @@ import ModelText from '@lblod/ember-rdfa-editor/model/model-text';
 import { XmlNodeRegistry } from '@lblod/ember-rdfa-editor/model/readers/xml-reader';
 import { compatTextAttributeMap } from '@lblod/ember-rdfa-editor/model/util/constants';
 import { TextAttribute } from '@lblod/ember-rdfa-editor/commands/text-properties/set-property-command';
-import { Mark } from '@lblod/ember-rdfa-editor/model/markSpec';
+import { Mark } from '@lblod/ember-rdfa-editor/model/mark';
 
 export default class XmlTextReader implements Reader<Element, ModelText, void> {
   constructor(private registry: XmlNodeRegistry<ModelText>) {}
