+++ addon/utils/ce/pernet-raw-editor.ts
@@ -43,7 +43,10 @@ import { tracked } from '@glimmer/tracking';
 import { Editor } from '@lblod/ember-rdfa-editor/editor/input-handlers/manipulation';
 import { ModelError } from '@lblod/ember-rdfa-editor/utils/errors';
 import { Region } from '@lblod/marawa/rdfa-block';
-import { INVISIBLE_SPACE } from '@lblod/ember-rdfa-editor/model/util/constants';
+import {
+  CORE_OWNER,
+  INVISIBLE_SPACE,
+} from '@lblod/ember-rdfa-editor/model/util/constants';
 import { ContentChangedEvent } from '@lblod/ember-rdfa-editor/utils/editor-event';
 
 export interface ContentObserver {
@@ -307,7 +310,14 @@ export default class PernetRawEditor extends RawEditor implements Editor {
         observer.handleFullContentUpdate(extraInfo);
       }
     }
-    this.eventBus.emit(new ContentChangedEvent());
+    this.eventBus.emit(
+      new ContentChangedEvent({
+        owner: CORE_OWNER,
+        payload: {
+          type: 'legacy',
+        },
+      })
+    );
   }
 
   /**
