+++ tests/unit/model/model-text-test.ts
@@ -1,6 +1,6 @@
 import { module, test } from 'qunit';
 import ModelText from '@lblod/ember-rdfa-editor/model/model-text';
-import { boldMarkSpec, Mark } from '@lblod/ember-rdfa-editor/model/markSpec';
+import { boldMarkSpec, Mark } from '@lblod/ember-rdfa-editor/model/mark';
 
 module('Unit | model | model-text-test', () => {
   module('Unit | model | model-text-test | sameAs', () => {
